Rename forms module to form
All other modules are in singular form, so change it for consistency.
