+++ crates/krilla/src/document.rs
@@ -17,7 +17,7 @@
 use crate::chunk_container::ChunkContainer;
 use crate::destination::NamedDestination;
 use crate::error::KrillaResult;
-use crate::forms::FieldTree;
+use crate::form::FieldTree;
 use crate::interchange::embed::EmbeddedFile;
 use crate::interchange::metadata::Metadata;
 use crate::interchange::outline::Outline;
