+++ crates/krilla/src/interactive/mod.rs
@@ -1,4 +1,4 @@
 pub mod action;
 pub mod annotation;
 pub mod destination;
-pub mod forms;
+pub mod form;
